pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = "A template for quickly getting started with django projects"
authors = ["SphericalKat <amolele@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
djangorestframework = "^3.12.1"
drf-spectacular = "^0.10.0"
gunicorn = "^20.0.4"
uvicorn = {extras = ["standard"], version = "^0.12.2"}
django-cors-headers = "^3.5.0"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
